treewide: scn[lib] everywhere pt VI: atoi()
Convert one of the two main cringe sources: atoi(),
to scn::scan_int<>() with strong result checking.
Also, convert ini::r_float() to scn::scan_value<f32>().

That was fun to fix configs and scripts, lemme tell you.
atoi() doesn't check the source (just returns 0 if it couldn't find
an integer), doesn't report errors. You could pass an empty buffer,
a signed value if you want to read unsigned etc.

dream.script even had r_float() for values which are strings in
dream.ltx. As a result, atof() always returned 0 and the actual
types were never parsed/used.

Next step is global atof() carnage.

+++ ogsr_engine/xrGame/damage_manager.cpp
@@ -9,6 +9,7 @@
 #include "stdafx.h"
 
 #include "damage_manager.h"
+
 #include "..\xr_3da\xr_object.h"
 #include "../Include/xrRender/Kinematics.h"
 #include "../xr_3da/bone.h"
@@ -88,7 +89,7 @@ void CDamageManager::load_section(LPCSTR section, CInifile* ini)
                        m_object->cNameVisual().c_str());
             CBoneInstance& bone_instance = kinematics->LL_GetBoneInstance(u16(bone));
             bone_instance.set_param(0, (float)atof(_GetItem(i.second.c_str(), 0, buffer)));
-            bone_instance.set_param(1, (float)atoi(_GetItem(i.second.c_str(), 1, buffer)));
+            bone_instance.set_param(1, (float)atof(_GetItem(i.second.c_str(), 1, buffer)));
             bone_instance.set_param(2, (float)atof(_GetItem(i.second.c_str(), 2, buffer)));
             if (_GetItemCount(i.second.c_str()) < 4)
             {
